requirements.txt
numpy==1.20.3
streamlit==0.84.0
wordcloud==1.8.1
matplotlib==3.2.2
pandas==1.3.2
contractions==0.0.52
nltk==3.6.1
textblob==0.17.1
seaborn==0.11.0
Pillow==8.4.0
scikit_learn==0.24.2
